requirements.txt
langchain~=0.0.265
bitsandbytes
accelerate
transformers~=4.30.2
datasets
loralib
sentencepiece
pypdf
sentence_transformers
openai
tiktoken
unstructured
tokenizers
faiss-cpu
xformers
pinecone-client
torch~=2.0.1
nltk~=3.8.1
typing~=3.7.4.3
streamlit~=1.25.0
fastapi~=0.100.1
pydantic~=1.10.9
requests~=2.31.0